refactor : Popup > Modal 네이밍 수정
