+++ src/components/Modal/index.ts
@@ -1,1 +1,1 @@
-export { default as Popup } from './Popup';
+export { default as Popup } from './Modal';
